refactor: Update unmarshaling and resolve process
Use interface{} across the unmarshaling and resolve processes.

BREAKING CHANGES:
- RenderingHandler.RenderYaml now returns interface{} instead of []byte
- Remove AnyObject.Value
- Add AnyObject.NormalizedValue
    returns normalized value with go built-in types
