+++ any.go
@@ -18,6 +18,20 @@ type AnyObjectMap struct {
 	Data map[string]*AnyObject `rs:"other"`
 }
 
+// NormalizedValue returns value of the AnyObjectMap with type map[string]interface{}
+func (o *AnyObjectMap) NormalizedValue() interface{} {
+	if o.Data == nil {
+		return map[string]interface{}(nil)
+	}
+
+	ret := make(map[string]interface{}, len(o.Data))
+	for k, v := range o.Data {
+		ret[k] = v.NormalizedValue()
+	}
+
+	return ret
+}
+
 func (md *AnyObjectMap) MarshalYAML() (interface{}, error) { return md.Data, nil }
 func (md *AnyObjectMap) MarshalJSON() ([]byte, error)      { return json.Marshal(md.Data) }
 
@@ -32,19 +46,30 @@ type AnyObject struct {
 	scalarData interface{}
 }
 
-func (o *AnyObject) Value() interface{} {
+// NormalizedValue returns value with primitive types of the AnyObject
+// that is:
+// 	for maps: map[string]interface{}
+// 	for slices: []interface{}
+//  and primitive types for scalar types
+//
+// should be called after fields being resolved
+func (o *AnyObject) NormalizedValue() interface{} {
 	switch {
 	case o.mapData != nil:
-		return o.mapData
+		return o.mapData.NormalizedValue()
 	case o.sliceData != nil:
-		return o.sliceData
+		ret := make([]interface{}, len(o.sliceData))
+		for i := range o.sliceData {
+			ret[i] = o.sliceData[i].NormalizedValue()
+		}
+		return ret
 	default:
 		return o.scalarData
 	}
 }
 
-func (o *AnyObject) MarshalYAML() (interface{}, error) { return o.Value(), nil }
-func (o *AnyObject) MarshalJSON() ([]byte, error)      { return json.Marshal(o.Value()) }
+func (o *AnyObject) MarshalYAML() (interface{}, error) { return o.NormalizedValue(), nil }
+func (o *AnyObject) MarshalJSON() ([]byte, error)      { return json.Marshal(o.NormalizedValue()) }
 
 func (o *AnyObject) UnmarshalYAML(n *yaml.Node) error {
 	switch n.Kind {
