+++ errors.go
@@ -1,0 +1,9 @@
+package rs
+
+import (
+	"errors"
+)
+
+var (
+	ErrInterfaceTypeNotHandled = errors.New("interface type not handled")
+)
